git/python-npr-to-video-project/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
moviepy>=1.0.3
deep-translator>=1.11.4
Pillow>=10.0.0
numpy>=1.24.0
